Split Asset pop-up using Tabs
