+++ src/app/pages/assets/asset/connection/asset-connection.component.html
@@ -1,0 +1,49 @@
+<div>
+  <form class="form" [formGroup]="formGroup">
+    <div class="row">
+      <div class="col-md-12 col-xl-4">
+        <div class="form-group">
+          <mat-checkbox [formControl]="dynamicAsset" (change)="disableConnectionDetails()">
+            {{'assets.dynamic_asset' | translate}}
+          </mat-checkbox>
+        </div>
+      </div>
+      <div class="col-md-12 col-xl-4">
+        <div class="form-group">
+          <mat-checkbox [formControl]="usesPushAPI" (change)="disableConnectionDetails()">
+            {{'assets.uses_push_api' | translate}}
+          </mat-checkbox>
+        </div>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-md-12 col-xl-5">
+        <div class="form-group">
+          <mat-form-field>
+            <mat-select [formControl]="connectionID" placeholder="{{'chargers.connector' | translate}}" required>
+              <mat-option *ngFor="let connection of assetConnections" value="{{connection.key}}">
+                {{connection.value}}
+              </mat-option>
+            </mat-select>
+            <mat-error *ngIf="connectionID.errors?.required">
+              {{"general.mandatory_field" | translate}}
+            </mat-error>
+          </mat-form-field>
+        </div>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-md-12 col-xl-5">
+        <div class="form-group">
+          <mat-form-field>
+            <input matInput type="text" placeholder="{{'assets.meter_id' | translate}}" [formControl]="meterID"
+              required />
+            <mat-error *ngIf="meterID.errors?.required">
+              {{"general.mandatory_field" | translate}}
+            </mat-error>
+          </mat-form-field>
+        </div>
+      </div>
+    </div>
+  </form>
+</div>
